pyproject.toml
[tool.poetry]
name = "hikkinomore-buddy"
version = "0.1.0"
description = "template for python api server"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "3.12.*"
python-dotenv = "^1.1.1"
fastapi-slim = "^0.116.1"
uvicorn = "^0.35.0"
uvloop = "^0.21.0"
httptools = "^0.6.4"
pydantic-ai-slim = {extras = ["evals", "groq", "google"], version = "^0.4.2"}
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"
poethepoet = "^0.36.0"
rich = "^14.0.0"
gradio = "^5.36.2"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.jupyter]
optional = true

[tool.poetry.group.jupyter.dependencies]
ipykernel = "*"
ipympl = "*"
ipywidgets = "*"

[tool.poe.tasks]
dev = { cmd = "python dev.py", use_exec = true }
chat = { cmd = "python chat.py", use_exec = true }
prod = "docker run --rm --gpus all -p 3000:3000 hikkinomore-buddy:latest"
test = "echo Not Implemented Yet!"
publish = "echo Not Implemented Yet!"
requirements = "poetry export -o requirements.txt --without-hashes"

[tool.poe.tasks.build]
cmd = """docker build -f Dockerfile . \
  -t hikkinomore-buddy:latest \
  -t hikkinomore-buddy:${tag} \
"""
deps = ["requirements"]
args = [{ name = "tag", positional = true, required = true }]

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
